+++ setupext/__init__.py
@@ -16,10 +16,8 @@
 #
 # *****************************************************************************
 
-from . import utils
-from . import dist
 from . import platform
 from . import build_ext
+from . import develop
 from . import test_java
 from . import sdist
-from . import pytester