+++ doc/CHANGELOG.rst
@@ -4,6 +4,38 @@ Changelog
 This changelog *only* contains changes from the *first* pypi release (0.5.4.3) onwards.
 
 Latest Changes:
+
+- **1.5.1_dev0 - 2023-12-15**
+
+  - Allow access to default methods implemented in interfaces when using ``@JImplements``.
+
+  - Added support for typing ``JArray`` (Java type only), e.g. ``JArray[java.lang.Object]`` ``"JArray[java.lang.Object]"``
+
+  - Fixed uncaught exception while setting traceback causing issues in Python 3.11/3.12.
+
+  - Use PEP-518 and PEP-660 configuration for the package, allowing editable and
+    configurable builds using modern Python packaging tooling.
+    Where before ``python setup.py --enable-tracing develop``, now can be done with
+    ``pip install --editable ./ --config-setting="--install-option=--enable-tracing"``.
+    The old setup.py usage remains, but is discouraged, and the arguments are now passed
+    after the command (previously they were specified before).
+
+  - Use PEP-518 configuration for the package, allowing
+    configurable builds using more up-to-date Python packaging tooling.
+    For editable installs, ``python setup.py --enable-tracing develop``
+    must now be done with ``python setup.py develop --enable-tracing``.
+
+  - Update for tests for numpy 2.0.
+
+  - Support of np.float16 conversion with arrays.
+
+  - Fixed a problem that caused ``dir(jpype.JPackage("mypackage"))`` to fail if
+    the class path contained non-ascii characters. See issue #1194.
+
+  - Fixed ``BufferOverflowException`` in ``JUnpickler`` when decoding
+    multiple Java objects.
+
+
 - **1.5.0 - 2023-04-03**
 
   - Support for Python 3.12
@@ -16,6 +48,8 @@ Latest Changes:
 
   - Java exceptions that occur in inequality comparisons now map to Python
     TypeError.
+
+- **1.4.2_dev0 - 2022-10-26**
 
   - Fixed crash when calling subscript on JArray.